Separated channel info form to another component
